move dependencies and metadata
