+++ geoparser/annotator/metadata.py
@@ -1,0 +1,26 @@
+tags_metadata = [
+    {
+        "name": "pages",
+        "description": "Navigation for different pages. Returns HTML templates",
+    },
+    {
+        "name": "session",
+        "description": "Management of user sessions.",
+    },
+    {
+        "name": "document",
+        "description": "Management of documents (as in parts of a session or an uploaded files).",
+    },
+    {
+        "name": "candidates",
+        "description": "Candidates for a specific toponym.",
+    },
+    {
+        "name": "annotation",
+        "description": "Toponym annotations.",
+    },
+    {
+        "name": "settings",
+        "description": "Settings on a session level.",
+    },
+]
